[MTDSA-10092] rename annual summary -> annual submission for retrieve uk prop models
