+++ app/v2/models/response/retrieveUkPropertyAnnualSubmission/ukNonFhlProperty/UkNonFhlProperty.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2021 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package v2.models.response.retrieveUkPropertyAnnualSubmission.ukNonFhlProperty
+
+import play.api.libs.functional.syntax._
+import play.api.libs.json._
+
+case class UkNonFhlProperty(adjustments: Option[UkNonFhlPropertyAdjustments], allowances: Option[UkNonFhlPropertyAllowances])
+
+object UkNonFhlProperty {
+  implicit val writes: OWrites[UkNonFhlProperty] = Json.writes[UkNonFhlProperty]
+
+  implicit val reads: Reads[UkNonFhlProperty] = (
+    (__ \ "ukOtherPropertyAnnualAdjustments").readNullable[UkNonFhlPropertyAdjustments] and
+      (__ \ "ukOtherPropertyAnnualAllowances").readNullable[UkNonFhlPropertyAllowances]
+  )(UkNonFhlProperty.apply _)
+}
