--- app/v2/models/response/retrieveUkPropertyAnnualSummary/ukNonFhlProperty/UkNonFhlPropertyBuilding.scala
@@ -1,32 +1,0 @@
-/*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package v2.models.response.retrieveUkPropertyAnnualSummary.ukNonFhlProperty
-
-import play.api.libs.functional.syntax._
-import play.api.libs.json._
-
-case class UkNonFhlPropertyBuilding(name: Option[String], number: Option[String], postcode: String)
-
-object UkNonFhlPropertyBuilding {
-  implicit val writes: OWrites[UkNonFhlPropertyBuilding] = Json.writes[UkNonFhlPropertyBuilding]
-
-  implicit val reads: Reads[UkNonFhlPropertyBuilding] = (
-    (__ \ "name").readNullable[String] and
-      (__ \ "number").readNullable[String] and
-      (__ \ "postCode").read[String]
-  )(UkNonFhlPropertyBuilding.apply _)
-}
